chore: move connection previewer out of subfolder
